+++ CMS/Models/PaginaApi.cs
@@ -3,7 +3,7 @@
     public class PaginaApi
 
     {
-        public int id { get; set; }
+        public ulong id { get; set; }
         public string Html { get; set; }        
     }
 }
