+++ CMS/Models/Repository/RepositoryDiv.cs
@@ -119,7 +119,7 @@ namespace CMS.Models.Repository
 
                 try
                 {
-                    ele = await contexto.Elemento.FirstOrDefaultAsync(d => d.Id == int.Parse(id));
+                    ele = await contexto.Elemento.FirstOrDefaultAsync(d => d.Id == ulong.Parse(id));
                     if (ele != null)
                     {
                         var paginaElementoDepe = contexto.Pagina.First(p => p.Id == ele.Pagina_);
@@ -231,7 +231,7 @@ namespace CMS.Models.Repository
                 div = await contexto.Div
                      .Include(d => d.Elemento)
                      .ThenInclude(d => d.Elemento)
-                     .FirstOrDefaultAsync(d => d.Id == int.Parse(id));
+                     .FirstOrDefaultAsync(d => d.Id == ulong.Parse(id));
 
                 if(div != null)
                 {
