+++ business/Join/CarouselImagem.cs
@@ -6,8 +6,8 @@ namespace business.Join
 {
     public class CarouselImagem
     {
-        public int? CarouselId { get; set; }
-        public int? ImagemId { get; set; }
+        public ulong? CarouselId { get; set; }
+        public ulong? ImagemId { get; set; }
         public CarouselImg Carousel { get; set; }
         public Imagem Imagem { get; set; }
     }
