+++ business/Join/DivPagina.cs
@@ -5,8 +5,8 @@ namespace business.Join
 {
     public class DivPagina
     {
-        public int? PaginaId { get; set; }
-        public int? DivId { get; set; }
+        public ulong? PaginaId { get; set; }
+        public ulong? DivId { get; set; }
         public virtual Div Div { get; set; }
         public virtual Pagina Pagina { get; set; }
     }
