+++ business/business/MensagemChat.cs
@@ -2,7 +2,7 @@
 {
     public class MensagemChat : BaseModel
     {
-        public int Pagina { get; set; }
+        public ulong Pagina { get; set; }
         public string NomeUsuario { get; set; }
         public string Mensagem { get; set; }
     }
